@@ -1,5 +1,5 @@
 {
-  "name": "basketstats-angular-frontend",
+  "name": "basketstats-app",
   "version": "0.0.1",
   "scripts": {
     "ng": "ng",
@@ -19,15 +19,22 @@
     "@angular/core": "^14.2.4",
     "@angular/forms": "^14.2.4",
     "@angular/localize": "^14.2.4",
-    "@angular/material": "^14.2.3",
+    "@angular/material": "^14.2.7",
     "@angular/platform-browser": "^14.2.4",
     "@angular/platform-browser-dynamic": "^14.2.4",
     "@angular/router": "^14.2.4",
     "@auth0/auth0-angular": "^1.11.0",
-    "@ng-bootstrap/ng-bootstrap": "^13.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^13.1.1",
+    "@ngrx/effects": "^14.3.3",
+    "@ngrx/entity": "^14.3.3",
+    "@ngrx/router-store": "^14.3.3",
+    "@ngrx/store": "^14.3.3",
+    "@ngrx/store-devtools": "^14.3.3",
+    "@popperjs/core": "^2.11.7",
     "@rxweb/reactive-form-validators": "^2.1.7",
     "bootstrap": "^4.5.2",
     "chart.js": "^2.9.4",
+    "date-fns": "^2.30.0",
     "eva-icons": "^1.1.2",
     "faker": "^5.2.0",
     "jquery": "^3.5.1",
@@ -38,14 +45,22 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.4",
-    "@angular/cli": "^14.2.4",
+    "@angular-devkit/build-angular": "^14.2.7",
+    "@angular-eslint/builder": "14.4.0",
+    "@angular-eslint/eslint-plugin": "14.4.0",
+    "@angular-eslint/eslint-plugin-template": "14.4.0",
+    "@angular-eslint/schematics": "14.4.0",
+    "@angular-eslint/template-parser": "14.4.0",
+    "@angular/cli": "^14.2.7",
     "@angular/compiler-cli": "^14.2.4",
-    "@schematics/angular": "^10.0.8",
+    "@ngrx/schematics": "^14.2.7",
+    "@schematics/angular": "^14.2.7",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "eslint": "~8.24.0",
+    "@typescript-eslint/eslint-plugin": "5.43.0",
+    "@typescript-eslint/parser": "5.43.0",
+    "eslint": "^8.28.0",
     "jasmine-core": "~4.3.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.4.1",
@@ -56,6 +71,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslib": "^2.4.0",
-    "typescript": "~4.6.4"
+    "typescript": "~4.8.2"
   }
 }
