@@ -1,16 +1,13 @@
-import {Component, OnInit, ViewChild} from '@angular/core';
-import {League} from "../../../models/league";
-import {Team} from "../../../models/team";
-import {TeamService} from "../../team/team.service";
-import {LeagueService} from "../../league/league.service";
-import {delay, first} from "rxjs/operators";
-import {User} from "../../../models/user";
-import {MatMenuTrigger} from "@angular/material/menu";
-import {UserAddEditComponent} from "../user-add-edit/user-add-edit.component";
-import {NgbModal} from "@ng-bootstrap/ng-bootstrap";
-import {LeagueAddEditComponent} from "./league-add-edit/league-add-edit.component";
-import {UserDeleteConfirmComponent} from "../user-delete-confirm/user-delete-confirm.component";
-import {LeagueDeleteConfirmComponent} from "./league-delete-confirm/league-delete-confirm.component";
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { MatMenuTrigger } from "@angular/material/menu";
+import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
+import { delay } from "rxjs/operators";
+import { League } from "../../../models/league.model";
+import { Team } from "../../../models/team";
+import { LeagueService } from "../../league/league.service";
+import { TeamService } from "../../team/team.service";
+import { LeagueAddEditComponent } from "./league-add-edit/league-add-edit.component";
+import { LeagueDeleteConfirmComponent } from "./league-delete-confirm/league-delete-confirm.component";
 
 @Component({
   selector: 'app-team-administration',
@@ -49,23 +46,23 @@ export class TeamAdministrationComponent implements OnInit {
 
   loadLeagues() {
     this.isLoading = true;
-    this.leagueService.getLeagueList().pipe(delay(100)).subscribe(data => {
+    this.leagueService.getAll().pipe(delay(100)).subscribe(data => {
       this.leagueList = data;
     });
   }
 
   getTeamsOfLeague(leagueId: number) {
-    return this.teamList? this.teamList.filter(obj => obj.league.id == leagueId): null;
+    return this.teamList ? this.teamList.filter(obj => obj.league.id == leagueId) : null;
   }
 
   addLeague() {
     const modalRef = this.modalService.open(LeagueAddEditComponent);
     modalRef.componentInstance.isEdit = false;
-    modalRef.result.then( res => {
+    modalRef.result.then(res => {
       if (res == 'confirm') {
         this.loadData();
       }
-    }, error =>{
+    }, error => {
 
     });
   }
@@ -74,11 +71,11 @@ export class TeamAdministrationComponent implements OnInit {
     const modalRef = this.modalService.open(LeagueAddEditComponent);
     modalRef.componentInstance.isEdit = true;
     modalRef.componentInstance.league = league;
-    modalRef.result.then( res => {
+    modalRef.result.then(res => {
       if (res == 'confirm') {
         this.loadData();
       }
-    }, error =>{
+    }, error => {
 
     });
   }
@@ -87,12 +84,12 @@ export class TeamAdministrationComponent implements OnInit {
     const modalRef = this.modalService.open(LeagueDeleteConfirmComponent);
     modalRef.componentInstance.leagueName = leagueName;
     modalRef.componentInstance.leagueId = leagueId;
-    modalRef.result.then( res => {
+    modalRef.result.then(res => {
       if (res == 'confirm') {
         this.loadData();
         this.loadData();
       }
-    }, error =>{
+    }, error => {
 
     });
   }
@@ -102,7 +99,7 @@ export class TeamAdministrationComponent implements OnInit {
     this.contextMenuPosition.x = event.clientX + 'px';
     this.contextMenuPosition.y = event.clientY + 'px';
     this.contextMenu.menuData = { 'league': league };
-    this.contextMenu.menu.focusFirstItem('mouse');
+    this.contextMenu.menu?.focusFirstItem('mouse');
     this.contextMenu.openMenu();
   }
 }
