@@ -1,11 +1,8 @@
 import { BreakpointObserver, Breakpoints } from "@angular/cdk/layout";
-import { Component, EventEmitter, OnDestroy, OnInit, Output } from '@angular/core';
-import {
-  NavigationEnd, Router, RouterEvent, RoutesRecognized
-} from "@angular/router";
+import { Component, EventEmitter, OnDestroy, Output } from '@angular/core';
+import { Router } from "@angular/router";
 import { Observable, Subject } from "rxjs";
-import { filter, map } from "rxjs/operators";
-import { ViewType } from "../../../enums/view-type.enum";
+import { map } from "rxjs/operators";
 import { User } from "../../../models/user";
 import { UserService } from "../../administration/user-list/user.service";
 
@@ -14,7 +11,7 @@ import { UserService } from "../../administration/user-list/user.service";
   templateUrl: './nav-header.component.html',
   styleUrls: ['./nav-header.component.scss']
 })
-export class NavHeaderComponent implements OnInit, OnDestroy{
+export class NavHeaderComponent implements OnDestroy {
 
   @Output() sidebarEvent = new EventEmitter<void>();
 
@@ -22,9 +19,6 @@ export class NavHeaderComponent implements OnInit, OnDestroy{
   username: string;
   user: User;
   userFullname: string;
-  activeViewType: ViewType;
-  ViewTypeEnum = ViewType;
-  menuOpened = false;
   isLoggedIn = false;
 
   isXSmall$: Observable<boolean> = this.breakpointObserver.observe(Breakpoints.XSmall)
@@ -33,31 +27,22 @@ export class NavHeaderComponent implements OnInit, OnDestroy{
     );
 
   constructor(private breakpointObserver: BreakpointObserver,
-              public userService: UserService,
-              public router: Router) {
+    public userService: UserService,
+    public router: Router) {
   }
 
-  ngOnInit() {
-    this.router.events.pipe(
-        filter((event: RouterEvent) => event instanceof NavigationEnd)
-      ).subscribe(() => {
-      this.isAuthenticated();
-      });
-    this.getViewType();
-  }
+  // ngOnInit() {
+  //   this.router.events.pipe(
+  //       filter((event: RouterEvent) => event instanceof NavigationEnd)
+  //     ).subscribe(() => {
+  //     this.isAuthenticated();
+  //     });
+  // }
 
   isAuthenticated() {
     // this.authService.isAuthenticated().subscribe((result => {
     //   this.isLoggedIn = result;
     // }));
-  }
-
-  getViewType() {
-    this.router.events.subscribe((data) => {
-      if (data instanceof RoutesRecognized) {
-        this.activeViewType = data.state.root.firstChild.data.viewType;
-      }
-    });
   }
 
   ngOnDestroy() {
