@@ -1,24 +1,31 @@
 package reforged.marcin.krysiak.basketstats.controllers;
 
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import reforged.marcin.krysiak.basketstats.models.JwtBlacklist;
-import reforged.marcin.krysiak.basketstats.service.JwtBlacklistService;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import reforged.marcin.krysiak.basketstats.security.Message;
 
 
-@RequestMapping("/api")
 @RestController
-@CrossOrigin
+@RequestMapping("/api/messages")
 public class AuthController {
 
-    @Autowired
-    private JwtBlacklistService jwtBlacklistService;
+    @GetMapping(value = "/public")
+    public Message publicEndpoint() {
+        return new Message("All good. You DO NOT need to be authenticated to call /api/public.");
+    }
 
+    @GetMapping(value = "/protected")
+    public Message privateEndpoint() {
+        return new Message("All good. You can see this because you are Authenticated.");
+    }
 
-    @PostMapping("/auth/logout")
-    public ResponseEntity<JwtBlacklist> logout(@RequestHeader("Authorization") String token) {
-        return ResponseEntity.ok().body(this.jwtBlacklistService.addToBlacklist(token));
+    @GetMapping(value = "/admin")
+    @PreAuthorize("hasAuthority('read:admin-messages')")
+    public Message privateScopedEndpoint() {
+        return new Message("All good. You can see this because you are Authenticated with a Token granted the 'read:admin-messages' scope");
     }
 }
