@@ -3,7 +3,7 @@ package reforged.marcin.krysiak.basketstats.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
-import reforged.marcin.krysiak.basketstats.config.enums.MatchStatus;
+import reforged.marcin.krysiak.basketstats.enums.MatchStatus;
 import reforged.marcin.krysiak.basketstats.models.Team;
 import reforged.marcin.krysiak.basketstats.models.User;
 
