@@ -1,15 +1,19 @@
 package reforged.marcin.krysiak.basketstats.service;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import reforged.marcin.krysiak.basketstats.models.League;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface LeagueService {
-    List<League> getAllLeagues();
-    Optional<League> getTeamById(Long id);
-    Optional<League> getLeagueByNane(String name);
-    League createLeague(League league);
-    void updateLeague(Long id, League league);
-    void deleteLeague(Long id);
+    Page<League> getAllBySpec(Specification<League> spec, Pageable pageable);
+    List<League> getAll();
+    Optional<League> getById(Long id);
+    Optional<League> getByNane(String name);
+    League create(League league);
+    void update(Long id, League league);
+    void delete(Long id);
 }
