@@ -1,6 +1,9 @@
 package reforged.marcin.krysiak.basketstats.service.impl;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import reforged.marcin.krysiak.basketstats.models.League;
@@ -11,34 +14,40 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-@Transactional
+@RequiredArgsConstructor
 public class LeagueServiceImpl implements LeagueService {
 
-    @Autowired
-    LeagueRepository leagueRepository;
+    private final LeagueRepository leagueRepository;
 
     @Override
-    public List<League> getAllLeagues() {
+    public Page<League> getAllBySpec(Specification<League> spec, Pageable pageable) {
+        return leagueRepository.findAll(spec, pageable);
+    }
+
+    @Override
+    public List<League> getAll() {
         return leagueRepository.findAllByOrderByIdDesc();
     }
 
     @Override
-    public Optional<League> getTeamById(Long id) {
+    public Optional<League> getById(Long id) {
         return leagueRepository.findById(id);
     }
 
     @Override
-    public Optional<League> getLeagueByNane(String name) {
+    public Optional<League> getByNane(String name) {
         return leagueRepository.findByName(name);
     }
 
     @Override
-    public League createLeague(League league) {
+    @Transactional
+    public League create(League league) {
         return leagueRepository.save(league);
     }
 
     @Override
-    public void updateLeague(Long id, League league){
+    @Transactional
+    public void update(Long id, League league) {
         if (this.leagueRepository.findById(id).isPresent()) {
             League newLeague = this.leagueRepository.findById(id).get();
             newLeague.setName(league.getName());
@@ -51,7 +60,7 @@ public class LeagueServiceImpl implements LeagueService {
     }
 
     @Override
-    public void deleteLeague(Long id) {
+    public void delete(Long id) {
         leagueRepository.deleteById(id);
     }
 }
