@@ -1,10 +1,7 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
-import { AuthGuard } from "./components/auth/auth-guard.service";
-import { AuthComponent } from "./components/auth/auth.component";
-import { LoginComponent } from "./components/auth/login/login.component";
-import { LogoutComponent } from "./components/auth/logout/logout.component";
 import { HomepageComponent } from "./components/homepage/homepage.component";
+import { LeagueComponent } from "./components/league/league.component";
 import { MatchStatsComponent } from "./components/schedule/match-stats/match-stats.component";
 import { ScheduleComponent } from "./components/schedule/schedule.component";
 import { TeamPlayersComponent } from "./components/team/team-players/team-players.component";
@@ -19,13 +16,25 @@ export const appRoutes: Routes = [
   },
 
   {
+    path: 'leagues',
+    data: {
+      viewType: ViewType.LEAGUES
+    },
+    children: [
+      {
+        path: '', component: LeagueComponent,
+      }
+    ]
+  },
+
+  {
     path: 'teams',
     data: {
       viewType: ViewType.TEAMS
     },
-    children:[
+    children: [
       {
-        path: '',  component: TeamComponent,
+        path: '', component: TeamComponent,
       },
       {
         path: 'info/:id', component: TeamPlayersComponent
@@ -38,9 +47,9 @@ export const appRoutes: Routes = [
     data: {
       viewType: ViewType.SCHEDULE
     },
-    children:[
+    children: [
       {
-        path: '',  component: ScheduleComponent,
+        path: '', component: ScheduleComponent,
       },
       {
         path: 'match/:id', component: MatchStatsComponent
@@ -48,44 +57,44 @@ export const appRoutes: Routes = [
     ]
   },
 
-  {
-    path: 'auth',
-    component: AuthComponent,
-    data: {
-    viewType: ViewType.AUTH
-    },
-    children: [
-      {
-        path: '',
-        component: LoginComponent,
-        data: {
-          roles: ['ROLE_GUEST']
-        },
-      },
-      {
-        path: 'login',
-        component: LoginComponent,
-        data: {
-          roles: ['ROLE_GUEST']
-        },
-      },
-      {
-        path: 'logout',
-        component: LogoutComponent,
-        canActivate: [AuthGuard]
-      },
-      {
-        path: 'request-password',
-        redirectTo: 'login'
-        // component: NbRequestPasswordComponent,
-      },
-      {
-        path: 'reset-password',
-        redirectTo: 'login'
-        // component: NbResetPasswordComponent,
-      },
-    ],
-  },
+  // {
+  //   path: 'auth',
+  //   component: AuthComponent,
+  //   data: {
+  //     viewType: ViewType.AUTH
+  //   },
+  //   children: [
+  //     {
+  //       path: '',
+  //       component: LoginComponent,
+  //       data: {
+  //         roles: ['ROLE_GUEST']
+  //       },
+  //     },
+  //     {
+  //       path: 'login',
+  //       component: LoginComponent,
+  //       data: {
+  //         roles: ['ROLE_GUEST']
+  //       },
+  //     },
+  //     {
+  //       path: 'logout',
+  //       component: LogoutComponent,
+  //       canActivate: [AuthGuard]
+  //     },
+  //     {
+  //       path: 'request-password',
+  //       redirectTo: 'login'
+  //       // component: NbRequestPasswordComponent,
+  //     },
+  //     {
+  //       path: 'reset-password',
+  //       redirectTo: 'login'
+  //       // component: NbResetPasswordComponent,
+  //     },
+  //   ],
+  // },
   { path: '', redirectTo: '', pathMatch: 'full' },
   { path: '**', redirectTo: '' },
 ];
@@ -94,9 +103,9 @@ export const appRoutes: Routes = [
   imports: [
     RouterModule.forRoot(appRoutes,
       {
-    useHash: true,
-    relativeLinkResolution: 'legacy'
-})
+        useHash: true,
+        relativeLinkResolution: 'legacy'
+      })
   ],
   exports: [
     RouterModule
