@@ -1,7 +1,8 @@
 import { BreakpointObserver, Breakpoints } from "@angular/cdk/layout";
 import { Component, OnInit } from '@angular/core';
-import { Observable } from "rxjs";
+import { BehaviorSubject, Observable } from "rxjs";
 import { map } from "rxjs/operators";
+import { LoaderService } from "./shared/service/loader/loader.service";
 
 @Component({
   selector: 'app-root',
@@ -15,16 +16,18 @@ export class AppComponent implements OnInit {
       map(result => result.matches)
     );
 
+  isLoading$: BehaviorSubject<boolean>;
+
   constructor(
-    private breakpointObserver: BreakpointObserver
+    private breakpointObserver: BreakpointObserver,
+    public loaderService: LoaderService
   ) { }
+
+  ngOnInit(): void {
+    this.isLoading$ = this.loaderService.isLoading$;
+  }
 
   toogleSidebar() {
     // this.nbSidebarService.toggle(false);
-  }
-
-  ngOnInit(): void {
-    // this.nbAuthService.isAuthenticatedOrRefresh().subscribe(res => {
-    // });
   }
 }
