@@ -1,14 +1,14 @@
-import { Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges, ViewChild } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
 import { MatMenuTrigger } from "@angular/material/menu";
 import { MatPaginator, PageEvent } from "@angular/material/paginator";
 import { MatSelectChange } from "@angular/material/select";
-import { MatSort, Sort } from "@angular/material/sort";
+import { MatSort } from "@angular/material/sort";
 import { MatTableDataSource } from "@angular/material/table";
 import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
 import { BehaviorSubject } from "rxjs";
 import { environment } from "../../../../../environments/environment";
 import { MatchStatus } from "../../../../enums/match-status.enum";
-import { League } from "../../../../models/league";
+import { League } from "../../../../models/league.model";
 import { Match } from "../../../../models/match";
 import { User } from "../../../../models/user";
 import { MatchService } from "../../../schedule/match.service";
@@ -23,7 +23,7 @@ import { MatchEditComponent } from "../match-edit/match-edit.component";
   templateUrl: './match-administration-list.component.html',
   styleUrls: ['./match-administration-list.component.css']
 })
-export class MatchAdministrationListComponent implements OnChanges, OnInit {
+export class MatchAdministrationListComponent implements OnInit {
 
   @Input()
   league: League;
@@ -51,7 +51,7 @@ export class MatchAdministrationListComponent implements OnChanges, OnInit {
 
   contextMenuPosition = { x: '0px', y: '0px' };
 
-  apiUrl = environment.url;
+  apiUrl = environment.api.url;
 
   MatchStatus = MatchStatus;
 
@@ -71,42 +71,6 @@ export class MatchAdministrationListComponent implements OnChanges, OnInit {
     });
   }
 
-  ngOnChanges(changes: SimpleChanges) {
-  }
-
-  onSortData(sort: Sort) {
-    let data = this.matchList.slice(this.lowIndex, this.highIndex);
-    if (sort.active && sort.direction !== '') {
-      data = data.sort((a: Match, b: Match) => {
-        const isAsc = sort.direction === 'asc';
-        const isDesc = sort.direction === 'desc';
-        switch (sort.active) {
-          case 'teamA': return this.compare(a.teamA.name, b.teamA.name, isAsc);
-          case 'teamB': return this.compare(a.teamB.name, b.teamB.name, isAsc);
-          case 'matchDate': return this.compareDate(a.matchDate, b.matchDate, isAsc);
-          case 'place': return this.compare(a.place, b.place, isAsc);
-          case 'finished': return this.compare(a.matchStatus, b.matchStatus, isDesc) && this.compareDate(a.matchDate, b.matchDate, isAsc);
-          default: return 0;
-        }
-      });
-    }
-    this.dataSubject.next(data);
-  }
-
-  private compare(a, b, isAsc) {
-    return (a < b ? -1 : 1) * (isAsc ? 1 : -1);
-  }
-
-  public compareDate(date1: Date, date2: Date, isAsc): number
-  {
-    let d1 = new Date(date1);
-    let d2 = new Date(date2);
-    let same = d1.getTime() === d2.getTime();
-    if (same) return 0;
-    else if (d1 > d2) return 1  * (isAsc ? 1 : -1);
-    else if (d1 < d2) return -1  * (isAsc ? 1 : -1);
-  }
-
   pageChange(paginator: PageEvent) {
     this.lowIndex = (paginator.pageIndex * paginator.pageSize);
     this.highIndex = this.lowIndex + paginator.pageSize;
@@ -117,11 +81,11 @@ export class MatchAdministrationListComponent implements OnChanges, OnInit {
     const modalRef = this.modalService.open(MatchAddComponent);
     modalRef.componentInstance.league = this.league;
     modalRef.componentInstance.userList = this.userList;
-    modalRef.result.then( res => {
+    modalRef.result.then(res => {
       if (res == 'confirm') {
         this.emitLoadData.emit();
       }
-    }, error =>{
+    }, error => {
 
     });
   }
@@ -131,11 +95,11 @@ export class MatchAdministrationListComponent implements OnChanges, OnInit {
     modalRef.componentInstance.isEdit = true;
     modalRef.componentInstance.league = this.league;
     modalRef.componentInstance.match = match;
-    modalRef.result.then( res => {
+    modalRef.result.then(res => {
       if (res == 'confirm') {
         this.emitLoadData.emit();
       }
-    }, error =>{
+    }, error => {
 
     });
   }
@@ -145,11 +109,11 @@ export class MatchAdministrationListComponent implements OnChanges, OnInit {
     modalRef.componentInstance.league = this.league;
     modalRef.componentInstance.match = match;
     modalRef.componentInstance.matchId = match.id;
-    modalRef.result.then( res => {
+    modalRef.result.then(res => {
       if (res == 'confirm') {
         this.emitLoadData.emit();
       }
-    }, error =>{
+    }, error => {
 
     });
   }
@@ -159,14 +123,14 @@ export class MatchAdministrationListComponent implements OnChanges, OnInit {
     const modalRef = this.modalService.open(LinkUserConfirmComponent);
     modalRef.componentInstance.match = match;
     modalRef.componentInstance.user = user;
-    modalRef.result.then( res => {
-        this.emitLoadData.emit();
-    }, error =>{
+    modalRef.result.then(res => {
+      this.emitLoadData.emit();
+    }, error => {
 
     });
   }
 
-  showToast(message: string, title: string, status, preventDuplicates, position, duration) {
+  showToast(message: string, title: string, status: string, preventDuplicates: boolean, position: string, duration: number) {
     // this.nbToastrService.show(message, title,
     //   { status, preventDuplicates, position, duration });
   }
@@ -176,8 +140,12 @@ export class MatchAdministrationListComponent implements OnChanges, OnInit {
     this.contextMenuPosition.x = event.clientX + 'px';
     this.contextMenuPosition.y = event.clientY + 'px';
     this.contextMenu.menuData = { 'row': row };
-    this.contextMenu.menu.focusFirstItem('mouse');
+    this.contextMenu.menu?.focusFirstItem('mouse');
     this.contextMenu.openMenu();
+  }
+
+  onSortData(event: any) {
+
   }
 
 }
@@ -188,11 +156,7 @@ export class RecordsDataSource extends MatTableDataSource<Match> {
     super();
   }
 
-  connect(): BehaviorSubject<any[]> {
+  override connect(): BehaviorSubject<any[]> {
     return this.subject;
-  }
-
-  disconnect(): void {
-
   }
 }
