@@ -1,13 +1,13 @@
+import { HttpClient } from "@angular/common/http";
 import { Injectable } from '@angular/core';
-import {environment} from "../../../environments/environment";
-import {HttpClient} from "@angular/common/http";
+import { environment } from "../../../environments/environment";
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthService {
 
-  private baseApiUrl = `${environment.url}/auth`;
+  private baseApiUrl = `${environment.api.url}/auth`;
 
   constructor(private http: HttpClient) { }
 
