@@ -1,35 +1,28 @@
+import { HttpClient } from "@angular/common/http";
 import { Injectable } from '@angular/core';
-import {environment} from "../../../environments/environment";
-import {HttpClient} from "@angular/common/http";
-import {Observable} from "rxjs";
+import { Observable } from "rxjs";
+import { League } from "src/app/models/league.model";
+import { BaseCrudService } from "src/app/shared/service/base-table/base-crud.service";
+import { QueryParamsService } from "src/app/shared/service/query-params/query-params.service";
 
 @Injectable({
   providedIn: 'root'
 })
-export class LeagueService {
+export class LeagueService extends BaseCrudService<League> {
 
-  private baseApiUrl = `${environment.url}/league`;
-
-  constructor(private http: HttpClient) { }
-
-  getLeagueList(): Observable<any> {
-    return this.http.get(`${this.baseApiUrl}/list`);
+  constructor(
+    http: HttpClient,
+    queryParamsService: QueryParamsService
+  ) {
+    super(
+      http,
+      queryParamsService,
+      '/league'
+    );
   }
 
   isLeagueWithNameExisting(name: string): Observable<Object> {
-    return this.http.get(`${this.baseApiUrl}/isLeagueWithNameExisting/${name}`);
-  }
-
-  createLeague(league: Object): Observable<Object> {
-    return this.http.post(`${this.baseApiUrl}/create`, league);
-  }
-
-  updateLeague(id: number, value: any): Observable<Object> {
-    return this.http.put(`${this.baseApiUrl}/update/${id}`, value);
-  }
-
-  deleteLeague(id: number): Observable<any> {
-    return this.http.delete(`${this.baseApiUrl}/delete/${id}`);
+    return this.http.get(`${this.fullResourceUrl}/isLeagueWithNameExisting/${name}`);
   }
 
 }
