@@ -1,9 +1,9 @@
-import {Component, Input, OnChanges, SimpleChanges, ViewChild} from '@angular/core';
-import {MatTableDataSource} from "@angular/material/table";
-import {MatSort} from "@angular/material/sort";
-import {MatchQuarterStatsForTeam} from "../../../../../../models/match-quarter-for-team";
-import {PlayerStats} from "../../../../../../models/player-stats";
-import {environment} from "../../../../../../../environments/environment";
+import { Component, Input, OnChanges, SimpleChanges, ViewChild } from '@angular/core';
+import { MatSort } from "@angular/material/sort";
+import { MatTableDataSource } from "@angular/material/table";
+import { environment } from "../../../../../../../environments/environment";
+import { MatchQuarterStatsForTeam } from "../../../../../../models/match-quarter-for-team";
+import { PlayerStats } from "../../../../../../models/player-stats";
 
 @Component({
   selector: 'app-quarter-stats-for-players-by-team',
@@ -22,7 +22,7 @@ export class QuarterStatsForPlayersByTeamComponent implements OnChanges {
 
   @ViewChild(MatSort, { static: true }) sort: MatSort;
 
-  apiUrl = environment.url;
+  apiUrl = environment.api.url;
 
   constructor() { }
 
