@@ -1,12 +1,12 @@
-import { Component, EventEmitter, Input, OnInit, Output, SimpleChanges, ViewChild } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
 import { MatMenuTrigger } from "@angular/material/menu";
 import { MatPaginator, PageEvent } from "@angular/material/paginator";
-import { MatSort, Sort } from "@angular/material/sort";
+import { MatSort } from "@angular/material/sort";
 import { MatTableDataSource } from "@angular/material/table";
 import { BehaviorSubject } from "rxjs";
 import { environment } from "../../../../../environments/environment";
 import { MatchStatus } from "../../../../enums/match-status.enum";
-import { League } from "../../../../models/league";
+import { League } from "../../../../models/league.model";
 import { Match } from "../../../../models/match";
 import { User } from "../../../../models/user";
 import { UserService } from "../../../administration/user-list/user.service";
@@ -45,7 +45,7 @@ export class UserMatchesListComponent implements OnInit {
 
   contextMenuPosition = { x: '0px', y: '0px' };
 
-  apiUrl = environment.url;
+  apiUrl = environment.api.url;
 
   MatchStatus = MatchStatus;
 
@@ -62,12 +62,12 @@ export class UserMatchesListComponent implements OnInit {
       this.dataSubject.next(this.matchList.slice(this.lowIndex, this.highIndex));
       this.dataSource = new RecordsDataSource(this.dataSubject);
       this.dataSource.sort = this.sort;
-      if(localStorage.getItem('teamAStatsForMatchInProgress') != null) {
+      if (localStorage.getItem('teamAStatsForMatchInProgress') != null) {
         this.matchIsInProgress = true;
       }
 
       this.matchList.forEach(m => {
-        if(m.matchStatus != MatchStatus.DONE
+        if (m.matchStatus != MatchStatus.DONE
           && m.matchStatus != MatchStatus.PLANNED
           && m.matchStatus != MatchStatus.SETTINGS) {
           this.matchIsInProgress = true;
@@ -76,96 +76,70 @@ export class UserMatchesListComponent implements OnInit {
     });
   }
 
-  ngOnChanges(changes: SimpleChanges) {
-  }
-
-  onSortData(sort: Sort) {
-    let data = this.matchList.slice(this.lowIndex, this.highIndex);
-    if (sort.active && sort.direction !== '') {
-      data = data.sort((a: Match, b: Match) => {
-        const isAsc = sort.direction === 'asc';
-        const isDesc = sort.direction === 'desc';
-        switch (sort.active) {
-          case 'teamA': return this.compare(a.teamA.name, b.teamA.name, isAsc);
-          case 'teamB': return this.compare(a.teamB.name, b.teamB.name, isAsc);
-          case 'matchDate': return this.compareDate(a.matchDate, b.matchDate, isAsc);
-          case 'place': return this.compare(a.place, b.place, isAsc);
-          case 'finished': return this.compare(a.matchStatus, b.matchStatus, isDesc) && this.compareDate(a.matchDate, b.matchDate, isAsc);
-          default: return 0;
-        }
-      });
-    }
-    this.dataSubject.next(data);
-  }
-
-  private compare(a, b, isAsc) {
-    return (a < b ? -1 : 1) * (isAsc ? 1 : -1);
-  }
-
-  public compareDate(date1: Date, date2: Date, isAsc): number
-  {
-    let d1 = new Date(date1);
-    let d2 = new Date(date2);
-    let same = d1.getTime() === d2.getTime();
-    if (same) return 0;
-    else if (d1 > d2) return 1  * (isAsc ? 1 : -1);
-    else if (d1 < d2) return -1  * (isAsc ? 1 : -1);
-  }
-
   pageChange(paginator: PageEvent) {
     this.lowIndex = (paginator.pageIndex * paginator.pageSize);
     this.highIndex = this.lowIndex + paginator.pageSize;
     this.dataSubject.next(this.matchList.slice(this.lowIndex, this.highIndex));
   }
 
-  showToast(message: string, title: string, status, preventDuplicates, position, duration) {
-    // this.nbToastrService.show(message, title,
-    //   { status, preventDuplicates, position, duration });
-  }
+  // showToast(message: string, title: string, status: string, preventDuplicates: boolean, position: string, duration: number) {
+  // this.nbToastrService.show(message, title,
+  //   { status, preventDuplicates, position, duration });
+  // }
 
-  onContextMenu(event: MouseEvent, row: Match) {
-    event.preventDefault();
-    this.contextMenuPosition.x = event.clientX + 'px';
-    this.contextMenuPosition.y = event.clientY + 'px';
-    this.contextMenu.menuData = { 'row': row };
-    this.contextMenu.menu.focusFirstItem('mouse');
-    this.contextMenu.openMenu();
-  }
+  // onContextMenu(event: MouseEvent, row: Match) {
+  //   event.preventDefault();
+  //   this.contextMenuPosition.x = event.clientX + 'px';
+  //   this.contextMenuPosition.y = event.clientY + 'px';
+  //   this.contextMenu.menuData = { 'row': row };
+  //   this.contextMenu.menu.focusFirstItem('mouse');
+  //   this.contextMenu.openMenu();
+  // }
 
-  getMatchInProgressQuarter(match: Match) {
+  getMatchInProgressQuarter(match: Match): number {
     if (match.matchStatus == MatchStatus.FIRST_QUARTER) {
       return 1;
-    } else if (match.matchStatus == MatchStatus.SECOND_QUARTER) {
+    }
+    if (match.matchStatus == MatchStatus.SECOND_QUARTER) {
       return 2;
-    } else if (match.matchStatus == MatchStatus.THIRD_QUARTER) {
+    } if (match.matchStatus == MatchStatus.THIRD_QUARTER) {
       return 3;
-    } else if (match.matchStatus == MatchStatus.FOURTH_QUARTER) {
+    }
+    if (match.matchStatus == MatchStatus.FOURTH_QUARTER) {
       return 4;
-    } else if (match.matchStatus == MatchStatus.OVERTIME) {
+    }
+    if (match.matchStatus == MatchStatus.OVERTIME) {
       return 5;
     }
+    throw (`Wrong match status. Got ${match.matchStatus}, expected: FIRST_QUARTER, SECOND_QUARTER, THIRD_QUARTER, FOURTH_QUARTER, OVERTIME`);
   }
 
-  getRouteLinkToMatch(match: Match) {
-    if(match.matchStatus == MatchStatus.PLANNED) {
-      return ['/match', 'settings', match.id];
-    } else if(match.matchStatus != MatchStatus.DONE) {
-
-      this.matchService.getStatsOfMatch(match.id).subscribe(res => {
-        console.log(res);
-        // if (match.matchStatus == MatchStatus.FIRST_QUARTER) {
-        //   return `['/match', 'progress', match.id, 1]`;
-        // } else if (match.matchStatus == MatchStatus.SECOND_QUARTER) {
-        //   return `['/match', 'progress', match.id, 2]`;
-        // } else if (match.matchStatus == MatchStatus.THIRD_QUARTER) {
-        //   return `['/match', 'progress', match.id, 3]`;
-        // } else if (match.matchStatus == MatchStatus.FOURTH_QUARTER) {
-        //   return `['/match', 'progress', match.id, 4]`;
-        // } else if (match.matchStatus == MatchStatus.OVERTIME) {
-        //   return `['/match', 'progress', match.id, 5]`;
-        // }
-      });
+  getRouteLinkToMatch(match: Match): string[] {
+    if (match.matchStatus == MatchStatus.PLANNED) {
+      return ['/match', 'settings', match.id.toString()];
     }
+    if (match.matchStatus != MatchStatus.DONE) {
+      if (match.matchStatus == MatchStatus.FIRST_QUARTER) {
+        return ['/match', 'progress', match.id.toString(), '1'];
+      }
+      if (match.matchStatus == MatchStatus.SECOND_QUARTER) {
+        return ['/match', 'progress', match.id.toString(), '2'];
+      }
+      if (match.matchStatus == MatchStatus.THIRD_QUARTER) {
+        return ['/match', 'progress', match.id.toString(), '3'];
+      }
+      if (match.matchStatus == MatchStatus.FOURTH_QUARTER) {
+        return ['/match', 'progress', match.id.toString(), '4'];
+      }
+      if (match.matchStatus == MatchStatus.OVERTIME) {
+        return ['/match', 'progress', match.id.toString(), '5'];
+      }
+    }
+    throw (`Wrong match status. Got ${match.matchStatus}, expected: DONE, PLANNED`);
+  }
+
+  onSortData(event: any): void {
+
   }
 
 }
@@ -176,11 +150,7 @@ export class RecordsDataSource extends MatTableDataSource<Match> {
     super();
   }
 
-  connect(): BehaviorSubject<any[]> {
+  override connect(): BehaviorSubject<any[]> {
     return this.subject;
-  }
-
-  disconnect(): void {
-
   }
 }
