@@ -3,7 +3,7 @@ import { Pipe, PipeTransform } from '@angular/core';
 @Pipe({ name: 'reverse' })
 
 export class ReversePipe implements PipeTransform {
-  transform(value) {
+  transform(value: any) {
     return value.slice().reverse();
   }
 }
