@@ -1,4 +1,6 @@
 export const environment = {
   production: true,
-  url : 'http://localhost:8080/api'
+  api: {
+    url: 'http://localhost:8080/api'
+  },
 };
